MS-14: Rename each js component to index.js
