+++ src/Components/Applications/Modal/index.js
@@ -1,7 +1,7 @@
 import styles from './modal.module.css';
-import Form from '../Form/Form.js';
-import Update from '../Update/Update.js';
-import Remove from '../Remove/Remove.js';
+import Form from '../Form';
+import Update from '../Update';
+import Remove from '../Remove';
 const Modal = (props) => {
   if (props.show == false) {
     return null;
