+++ src/Components/Applications/Remove/index.js
